# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TraderMemos API
  version: 1.0.0
  description: |
    Self-hosted trading journal API.

    Authenticate with either:
    - A short-lived session JWT from `POST /api/v1/auth/login` (web app)
    - A personal access token (`tm_pat_…`) from Settings → API or `POST /api/v1/access-tokens`

    Send the secret as `Authorization: Bearer <token>`.
servers:
  - url: /
    description: This API host
...
paths:
...
  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      operationId: listAccounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Accounts]
      summary: Create account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/Error"
...
  /api/v1/analytics/summary:
    get:
      tags: [Analytics]
      summary: P&L summary
      description: Normalizes each closed trade's net P&L, gross P&L and fees to the output currency before recomputing statistics. Uses latest FX, not historical trade-date rates. Mixed account scopes require target_currency.
      operationId: analyticsSummary
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - name: target_currency
          in: query
          description: Three-letter target currency. When omitted, uses the shared account currency and rejects mixed scopes.
          schema:
            type: string
            pattern: '^[A-Za-z]{3}$'
      responses:
        "200":
          description: Recomputed summary in the actual output currency.
          content:
            application/json:
              schema:
                type: object
                required: [currency, target_currency, fx_policy, fx_rates]
                properties:
                  currency:
                    type: string
                    description: Actual output currency; empty only for an empty account scope without a requested target.
                  target_currency:
                    type: string
                    description: Requested target (uppercase); empty when omitted.
                  fx_policy:
                    type: string
                    enum: [latest]
                  fx_rates:
                    type: array
                    description: Rates applied once per source currency, including provider and as_of. Identity conversions need no FX lookup.
                    items:
                      type: object
                      additionalProperties: true
                additionalProperties: true
        "400":
          description: Invalid target, mixed_currencies without a target, or unknown_currency; no monetary aggregate is returned.
        "502":
          description: fx_unavailable; required FX is missing, invalid, or failed. No partial summary is returned.
  /api/v1/analytics/daily:
    get:
      tags: [Analytics]
      summary: Daily P&L
      operationId: analyticsDaily
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - $ref: "#/components/parameters/TzQuery"
        - name: target_currency
          in: query
          description: Three-letter target currency. When omitted, uses the shared account currency and rejects mixed scopes.
          schema:
            type: string
            pattern: '^[A-Za-z]{3}$'
      responses:
        "200":
          description: Daily map without target_currency (legacy clients); normalized envelope with an explicit target.
          content:
            application/json:
              schema:
                type: object
                description: With target_currency, contains currency, target_currency, fx_policy (latest), fx_rates and pnl (date to amount). Without target, returns the legacy date to amount map.
                additionalProperties: true
        "400":
          description: Invalid target, unknown currency or mixed scope without target.
        "502":
          description: fx_unavailable; no partial results are returned.
  /api/v1/analytics/behavior:
    get:
      tags: [Analytics]
      summary: Behavioral pattern report (revenge trading, overconfidence, loss aversion)
      operationId: analyticsBehavior
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - $ref: "#/components/parameters/TzQuery"
        - name: target_currency
          in: query
          description: Convert each monetary input before detecting patterns; latest FX and fail-closed.
          schema:
            type: string
      responses:
        "200":
          description: Behavior report with currency, target_currency, fx_policy latest and fx_rates
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/analytics/equity-curve:
    get:
      tags: [Analytics]
      summary: Equity curve
      operationId: analyticsEquityCurve
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - name: target_currency
          in: query
          description: Three-letter target currency. When omitted, uses the shared account currency and rejects mixed scopes.
          schema:
            type: string
            pattern: '^[A-Za-z]{3}$'
      responses:
        "200":
          description: Normalized equity points, monetary drawdown and cash_transactions in the actual output currency.
          content:
            application/json:
              schema:
                type: object
                required: [currency, target_currency, fx_policy, fx_rates]
                properties:
                  currency:
                    type: string
                    description: Actual output currency; empty only for an empty account scope without a requested target.
                  target_currency:
                    type: string
                    description: Requested target (uppercase); empty when omitted.
                  fx_policy:
                    type: string
                    enum: [latest]
                  fx_rates:
                    type: array
                    description: Rates applied once per source currency, including provider and as_of. Identity conversions need no FX lookup.
                    items:
                      type: object
                      additionalProperties: true
                additionalProperties: true
        "400":
          description: Invalid target, unknown currency or mixed scope without target.
        "502":
          description: fx_unavailable; no partial results are returned.
  /api/v1/analytics/account-value:
    get:
      tags: [Analytics]
      summary: Reconstructed daily account value
      operationId: analyticsAccountValue
      parameters:
        - $ref: "#/components/parameters/AccountIdQuery"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
      responses:
        "200":
          description: Daily mark-to-market account value
          content:
            application/json:
              schema:
                type: object
                required: [currency, timezone, adjustment_status, points]
                properties:
                  currency:
                    type: string
                  timezone:
                    type: string
                  adjustment_status:
                    type: string
                  points:
                    type: array
                    items:
                      type: object
                      required: [date, contributed_capital, cash_balance, realized_pnl, status, warnings]
                      properties:
                        date:
                          type: string
                          format: date
                        estimated_account_value:
                          type: [number, "null"]
                        contributed_capital:
                          type: number
                        cash_balance:
                          type: number
                        open_position_value:
                          type: [number, "null"]
                        realized_pnl:
                          type: number
                        unrealized_pnl:
                          type: [number, "null"]
                        status:
                          type: string
                        warnings:
                          type: array
                          items:
                            type: object
                            required: [code, date, message]
                            properties:
                              code:
                                type: string
                              instrument:
                                type: string
                              execution_id:
                                type: string
                              date:
                                type: string
                                format: date
                              message:
                                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session JWT access token **or** personal access token (`tm_pat_…`).
  parameters:
...
    AccountIdQuery:
      name: account_id
      in: query
      description: >-
        Account scope. Accepts a single account id, a comma-separated list, or
        the parameter repeated (portfolio mode). All included accounts must share
        one base currency; mixed-currency selections return 400
        `mixed_currencies`, including when omitted. Omit for all live accounts.
      schema:
        type: string
    FromQuery:
      name: from
      in: query
      description: RFC3339 start time
      schema:
        type: string
        format: date-time
    ToQuery:
      name: to
      in: query
      description: RFC3339 end time
      schema:
        type: string
        format: date-time
    TzQuery:
      name: tz
      in: query
      description: >
        IANA timezone for day / hour-of-day / weekday bucketing
        (e.g. `America/New_York`). Defaults to UTC. Session bucketing
        always uses the US-equity clock (Eastern).
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        broker:
          type: string
        account_type:
          type: string
          description: >-
            Free-form account kind — cash, margin, prop, or backtest. Accounts
            marked backtest are paper accounts fed by the bar replay; their
            trades and cash are excluded from all-account aggregates unless the
            account is selected explicitly via account_id.
        base_currency:
          type: string
        starting_balance:
          type: number
    CreateAccount:
      type: object
      required: [name]
      properties:
        name:
          type: string
        broker:
          type: string
        account_type:
          type: string
          description: >-
            Free-form account kind — cash, margin, prop, or backtest. Accounts
            marked backtest are paper accounts fed by the bar replay; their
            trades and cash are excluded from all-account aggregates unless the
            account is selected explicitly via account_id.
        base_currency:
          type: string
        starting_balance:
          type: number
...
security:
  - bearerAuth: []